feat: adding first FormGroup story
