+++ stories/11.FormGroupExplanation.stories.mdx
@@ -1,0 +1,111 @@
+import { Meta, Story, Canvas, Props } from '@storybook/addon-docs/blocks';
+import { FormGroup } from '../src/formGroup/FormGroup';
+import './style.css';
+
+<Meta title="DCXLibrary/Form/Group/documentation" component={FormGroup} />
+
+# FormGroup
+
+A group component ready to use in your form. 
+
+FormGroup is UI/UX agnostic so you need to provide your style to have the look and feel you prefer.
+
+## Usage
+
+Use the group component when users can only select one option from a list.
+
+It provides heading, hint text and error message components. 
+
+An example with all the available properties is:
+
+```js
+<FormGroup
+    name='name-of-group'
+    items={[
+        {
+            label: 'item-label',
+            value: 'yes',
+            ariaLabel: 'aria-label',
+            disabled: false,
+            hint: {
+                text: 'hint-text',
+                classes: 'class names',
+                id: 'id'
+            },
+            id: 'id',
+            inputProps: {
+                style: {
+                    lineHeight: '10px'
+                }
+            },
+            itemProps: {
+                style: {
+                    display: 'flex'
+                }
+            },
+            labelProps: {
+                style: {
+                    fontSize: '14px'
+                }
+            },
+            selected: false,
+        },
+        {
+            label: 'item-label',
+            value: 'no',
+            ariaLabel: 'aria-label',
+            disabled: false,
+            hint: {
+                text: 'hint-text',
+                classes: 'class names',
+                id: 'id'
+            },
+            id: 'id',
+            inputProps: {
+                style: {
+                    lineHeight: '10px'
+                }
+            },
+            itemProps: {
+                style: {
+                    display: 'flex'
+                }
+            },
+            labelProps: {
+                style: {
+                    fontSize: '14px'
+                }
+            },
+            selected: false
+        }
+    ]}
+    ariaDescribedBy='aria-described-by'
+    groupClassess='group-classes'
+    error={{
+       text: 'error message'
+       classes: 'error-classes'
+       id: 'error-id',
+       visuallyHiddenText: {
+           classes='visually-hidden-classes',
+           text='visually hidden text'
+       }
+    }}
+    fieldsetClasses='fieldset classes'
+    hint={
+        text: 'hint-text',
+        classes: 'class names',
+        id: 'hint-id'
+    }
+    id='id'
+    itemsClasses='class names for the group items'
+    legend={
+        text='heading',
+        classes='legend classes',
+        isHeading=true,
+        headingClasses='heading classes'
+    }
+    onChange={handleChange}
+/>
+```
+
+<Props of={FormGroup} />
